requirements.txt
discord.py
riotwatcher
psycopg2